meebot/requirements.txt
discord.py==2.4.0
aiosqlite==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
chromadb==0.5.23
aiohttp==3.10.11
cryptography==42.0.8
